Plotting gray values with origin at *heart* surface
After discussing the data with Ludovic I wanted to update the plots to
only show them up to 2mm depth.
Then it dawned on me that I've plotted the gray values with wrong
'coordinates' all the time, i.e. making them hard to compare (if not
impossible).

The origin of the detailed plots should be at the *heart* surface, not
the sample surface!
After a bit of trickery with flipping arrays and coordinates the plots
are *much* (*MUCH*) easier to compare now, and we also see a hint on the
patch thickness of 300 um.

This is the version where I've sent the plots again for discussion with
Ludovic.